Refactor BottomSheet and ButtonMenu
