+++ snap-geometry-edits/src/main/java/com/esri/arcgismaps/sample/snapgeometryedits/screens/MainScreen.kt
@@ -32,9 +32,7 @@ import com.arcgismaps.toolkit.geoviewcompose.MapView
 import com.esri.arcgismaps.sample.sampleslib.components.BottomSheet
 import com.esri.arcgismaps.sample.sampleslib.components.MessageDialog
 import com.esri.arcgismaps.sample.sampleslib.components.SampleTopAppBar
-import com.esri.arcgismaps.sample.snapgeometryedits.components.ButtonMenu
 import com.esri.arcgismaps.sample.snapgeometryedits.components.MapViewModel
-import com.esri.arcgismaps.sample.snapgeometryedits.components.SnapSettings
 
 /**
  * Main screen layout for the sample app
